solution 4

+++ src/routes/accounts.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 from typing import cast
-
+from sqlalchemy.ext import IntegrityError
 from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy import select, delete
 from sqlalchemy.exc import SQLAlchemyError
@@ -31,14 +31,16 @@ from schemas import (
     TokenRefreshRequestSchema
 )
 
+from src.exceptions.security import TokenExpiredError, InvalidTokenError
 from src.security.utils import generate_secure_token
 
 router = APIRouter()
 
 
 @router.post("/register/", status_code=201)
-async def register(user: UserRegistrationSchema, db: AsyncSession = Depends(get_db)):
+async def register(user: UserRegistrationRequrstSchema, db: AsyncSession = Depends(get_db), jwt_manager: JWTAuthManagerInterface = Depends(get_jwt_auth_manager)):
     db_user = await db.scalar(select(UserModel).where(UserModel.email == user.email))
+    access_token = jwt_manager.create_access_token(user.model_dump())
     group = await db.scalar(select(UserGroupModel).where(UserGroupModel.name == UserGroupEnum.USER))
     if db_user:
         raise HTTPException(status_code=409, detail=f"A user with this email {user.email} already exists.")
@@ -48,7 +50,7 @@ async def register(user: UserRegistrationSchema, db: AsyncSession = Depends(get_
         await db.flush()
         activation_token = ActivationTokenModel(
             token=access_token,
-            user_id=user.id
+            user_id=new_user.id
         )
         db.add(activation_token)
         await db.commit()
@@ -58,7 +60,7 @@ async def register(user: UserRegistrationSchema, db: AsyncSession = Depends(get_
         await db.rollback()
         raise HTTPException(
             status_code=409,
-            detail=f"A user with this email {user_data.email} already exists."
+            detail=f"A user with this email {user.email} already exists."
         )
     except Exception:
         await db.rollback()
@@ -78,6 +80,9 @@ async def activate(user: UserActivationSchema, db: AsyncSession = Depends(get_db
         raise HTTPException(status_code=400, detail="User account is already active.")
     if not db_user.activation_token:
         raise HTTPException(status_code=400, detail="Invalid or expired activation token.")
+    expires_at = db_user.activation_token.expires_at.replace(tzinfo=timezone.utc)
+    if db_user.activation_token.token != user.token or expires_at < datetime.now(tz=timezone.utc):
+        raise HTTPException(status_code=400, detail="Invalid or expired activation token.")
 
     db_user.is_active = True
     await db.delete(db_user.activation_token)
@@ -93,7 +98,7 @@ async def request_password_reset_token(user: UserBaseSchema, db: AsyncSession = 
         return MessageResponseSchema(message="If you are registered, you will receive an email with instructions.")
     if db_user.password_reset_token:
         await db.delete(user.password_reset_token)
-    user.password_reset_token = PasswordResetTokenModel(
+    db_user.password_reset_token = PasswordResetTokenModel(
         token=generate_secure_token(),
         user_id=db_user.id
     )
@@ -107,18 +112,18 @@ async def reset_password_complete(user: PasswordResetCompleteSchema, db: AsyncSe
 
     if not db_user or not db_user.password_reset_token:
         raise HTTPException(status_code=400, detail="Invalid email or token.")
-    expires_at = user.password_reset_token.expires_at.replace(tzinfo=timezone.utc)
-    if user.password_reset_token.token != complete_data.token or expires_at < datetime.now(tz=timezone.utc):
-        await db.delete(user.password_reset_token)
+    expires_at = db_user.password_reset_token.expires_at.replace(tzinfo=timezone.utc)
+    if db_user.password_reset_token.token != user.token or expires_at < datetime.now(tz=timezone.utc):
+        await db.delete(db_user.password_reset_token)
         await db.commit()
         raise HTTPException(status_code=400, detail="Invalid email or token.")
     try:
         db_user.password = hash_password(user.password)
-        await db.delete(user.password_reset_token)
+        await db.delete(db_user.password_reset_token)
         await db.commit()
         return MessageResponseSchema(message="Password reset complete.")
     except Exception:
-        db.rollback()
+        await db.rollback()
         raise HTTPException(status_code=500, detail="An error occurred while resetting the password.")
 
 
@@ -152,7 +157,7 @@ async def refresh_account_token(refresh_token: TokenRefreshRequestSchema,
                                 jwt_manager: JWTAuthManagerInterface = Depends(get_jwt_auth_manager),
                                 db: AsyncSession = Depends(get_db)):
     try:
-        jwt_manager.decode_refresh_token(refresh_token_data.refresh_token)
+        jwt_manager.decode_refresh_token(refresh_token.refresh_token)
     except TokenExpiredError:
         raise HTTPException(status_code=400, detail="Token has expired.")
     except InvalidTokenError:
